Khansole Academy
Khansole Academy is a program that helps other people learn addition. The program will generate a random simple addition problem that involves adding two 2-digit integers. The user is asked for an answer to the generated problem. The program determines if the answer is correct or not, and gives the user an appropriate message to let them know. The user needs to get 3 problems correctly in a row to complete the program.