Hailstones
Based on a problem in Douglas Hofstadter’s Pulitzer-prize-winning book Gödel, Escher, Bach. That book contains many interesting mathematical puzzles, many of which can be expressed in the form of computer programs. In Chapter XII, Hofstadter mentions a wonderful problem. The problem can be expressed as follows:

"Pick some positive integer and call it n.
If n is even, divide it by two.
If n is odd, multiply it by three and add one.
Continue this process until n is equal to one."

I want to write a program that tests every positive integer between 1 and 100.
It should give me:
+ number 
+ amount of steps to reach 1

EXTRA show the results in a scatter plot, using matplotlib.